+++ src/js/policies/firefox.js
@@ -406,8 +406,8 @@ const policies = {
         ]
       },
       {
-        name : 'Highlights',
-        label : browser.i18n.getMessage('policy_description_FirefoxHome_Highlights'),
+        name : 'SponsoredTopSites',
+        label : browser.i18n.getMessage('policy_description_FirefoxHome_SponsoredTopSites'),
         type : 'enum',
         options : [
           {
@@ -444,8 +444,27 @@ const policies = {
         ]
       },
       {
-        name : 'Snippets',
-        label : browser.i18n.getMessage('policy_description_FirefoxHome_Snippets'),
+        name : 'SponsoredPocket',
+        label : browser.i18n.getMessage('policy_description_FirefoxHome_SponsoredPocket'),
+        type : 'enum',
+        options : [
+          {
+            label : browser.i18n.getMessage('enum_value_no_preference'),
+            value : null
+          },
+          {
+            label : browser.i18n.getMessage('enum_value_yes'),
+            value : 'true'
+          },
+          {
+            label : browser.i18n.getMessage('enum_value_no'),
+            value : 'false'
+          }
+        ]
+      },
+      {
+        name : 'Highlights',
+        label : browser.i18n.getMessage('policy_description_FirefoxHome_Highlights'),
         type : 'enum',
         options : [
           {